+++ lightsim2grid/tests/test_init_from_pypowsybl.py
@@ -10,13 +10,10 @@ import pypowsybl as pp
 import pypowsybl.loadflow as lf
 import numpy as np
 import unittest
-import tempfile
-import os
 
 from lightsim2grid.gridmodel.from_pypowsybl import init
 
 import pandapower.networks as pn
-import pandapower.converter as pc
 import pandapower as pdp
 from lightsim2grid.gridmodel import init as init_from_pp
 
@@ -24,24 +21,25 @@ from lightsim2grid.gridmodel import init as init_from_pp
 class AuxInitFromPyPowSyBl:    
     def get_pypo_grid(self):
         return pp.network.create_ieee14()
-    def fix_grid_pypow(self):
-        # see https://github.com/powsybl/pypowsybl/issues/642
-        vn = 14.
-        self.network.update_lines(id=["L7-8-1", "L7-9-1"],
-                                  x=[0.17615 * vn*vn/100., 0.11001 * vn*vn/100.],
-                                  b1=[0., 0.],
-                                  b2=[0., 0.])
-        lf.run_ac(self.network)    
+         
     def get_gen_slack_id(self):
         return 0
+    
     def get_slackbus_id(self):
         return 0
+    
     def get_equiv_pdp_grid(self):
         return pn.case14()
     
+    def get_tol_eq(self):
+        return 1e-6
+    
+    def compare_pp(self):
+        """will this test suite compare pypowsybl and pandapower (cannot be used for ieee57 or ieee118)"""
+        return True
+    
     def setUp(self) -> None:
-        self.network = self.get_pypo_grid()
-        self.fix_grid_pypow()
+        self.network_ref = self.get_pypo_grid()
         
         # init equivalent pandapower grid (if any)
         tmp = self.get_equiv_pdp_grid()
@@ -56,64 +54,69 @@ class AuxInitFromPyPowSyBl:
             self.ref_samecase = None
             
         # init lightsim2grid model
-        self.gridmodel = init(self.network, gen_slack_id=self.get_gen_slack_id())
+        self.gridmodel = init(self.network_ref, gen_slack_id=self.get_gen_slack_id())
         
         # use some data
-        self.nb_bus_total = self.network.get_buses().shape[0]
+        self.nb_bus_total = self.network_ref.get_buses().shape[0]
         self.V_init_dc = np.ones(self.nb_bus_total, dtype=np.complex_)
         self.V_init_ac = 1.04 * self.V_init_dc
-        self.tol = 1e-7
-        self.tol_eq = 1e-6
+        self.tol = 1e-7  # for the solver
+        self.tol_eq = self.get_tol_eq()
         return super().setUp()
     
     def test_basic(self):
         """check that all elements are ok"""
-        assert len(self.gridmodel.get_lines()) == self.network.get_lines().shape[0]    
-        assert len(self.gridmodel.get_loads()) == self.network.get_loads().shape[0]    
-        assert len(self.gridmodel.get_generators()) == self.network.get_generators().shape[0]    
-        assert len(self.gridmodel.get_trafos()) == self.network.get_2_windings_transformers().shape[0]    
-        assert len(self.gridmodel.get_shunts()) == self.network.get_shunt_compensators().shape[0]    
+        assert len(self.gridmodel.get_lines()) == self.network_ref.get_lines().shape[0]    
+        assert len(self.gridmodel.get_loads()) == self.network_ref.get_loads().shape[0]    
+        assert len(self.gridmodel.get_generators()) == self.network_ref.get_generators().shape[0]    
+        assert len(self.gridmodel.get_trafos()) == self.network_ref.get_2_windings_transformers().shape[0]    
+        assert len(self.gridmodel.get_shunts()) == self.network_ref.get_shunt_compensators().shape[0]    
     
     def test_compare_pp(self):
         """compare from the reference case14"""
         if not self.can_pp:
             self.skipTest("no equivalent pandapower grid")
+        if not self.compare_pp():
+            self.skipTest("It is expected that pandapower and pypowsybl gives different results") 
             
-        # check the SBus and Ybus in DC (i need to run powerflow for that)
-        v_ls = self.gridmodel.dc_pf(self.V_init_dc, 2, self.tol)
-        v_ls_ref = self.ref_samecase.dc_pf(self.V_init_dc, 2, self.tol)
+        # check the SBus and Ybus in DC (i need to run powerflow for that)      
+        v_ls = self.gridmodel.dc_pf(1.0 * self.V_init_dc, 2, self.tol)
+        v_ls_ref = self.ref_samecase.dc_pf(1.0 * self.V_init_dc, 2, self.tol)
         slack_id = self.get_slackbus_id()
+        # (array([ 30,  31, 212, 218, 265]), array([265,  31, 212, 218,  30]))
         
-        bus_or = [67]
-        bus_ex = [115]
-        l_id_ls = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or) and (el.bus_ex_id in bus_ex) ]
-        self.gridmodel.get_lines()[173].r_pu
-        self.gridmodel.get_lines()[173].x_pu
-        self.gridmodel.get_lines()[173].h_pu
+        import pdb
+        pdb.set_trace()
         
-        bus_or = [85]
-        bus_ex = [86]
-        l_id_ls = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or) and (el.bus_ex_id in bus_ex) ]
-        self.gridmodel.get_lines()[l_id_ls[0]].r_pu
-        self.gridmodel.get_lines()[l_id_ls[0]].x_pu
-        self.gridmodel.get_lines()[l_id_ls[0]].h_pu
+        # for case 118
+        # bus_or = [64]
+        # bus_ex = [67]
+        # lines = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or and el.bus_ex_id in bus_ex) or (el.bus_or_id in bus_ex and el.bus_ex_id in bus_or)]
+        # lines_ref = [el.id for el in self.ref_samecase.get_lines() if (el.bus_or_id in bus_or and el.bus_ex_id in bus_ex) or (el.bus_or_id in bus_ex and el.bus_ex_id in bus_or)]
+        # if not lines_ref:
+        #     lines_ref = [el.id for el in self.ref_samecase.get_trafos() if (el.bus_hv_id in bus_or and el.bus_lv_id in bus_ex) or (el.bus_hv_id in bus_ex and el.bus_lv_id in bus_or)]
+        # import pdb
+        # pdb.set_trace()
+        # # self.pp_samecase["_ppc"]["internal"]["Bbus"]
+        # self.pp_samecase["_ppc"]["internal"]["Ybus"][64,67]
+        # self.pp_samecase["_ppc"]["internal"]["bus"]
+
+        # for case 300
+        # np.where(tmp_.todense() >= 2000)
+        # (array([ 30,  31, 212, 218, 265]), array([265,  31, 212, 218,  30]))
 
         bus_or = [64]
         bus_ex = [67]
-        l_id_ls = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or) and (el.bus_ex_id in bus_ex) ]
-        self.gridmodel.get_lines()[l_id_ls[0]].r_pu
-        self.gridmodel.get_lines()[l_id_ls[0]].x_pu
-        self.gridmodel.get_lines()[l_id_ls[0]].h_pu
-        
-        l_id_ls_ref = [el.id for el in self.ref_samecase.get_lines() if (el.bus_or_id in bus_or) and (el.bus_ex_id in bus_ex) ]
-        l_id_ls_ref = [el.id for el in self.ref_samecase.get_trafos() if (el.bus_hv_id in bus_or) and (el.bus_lv_id in bus_ex) ]
-        self.ref_samecase.get_lines()[l_id_ls_ref[0]].r_pu
-        self.ref_samecase.get_lines()[l_id_ls_ref[0]].x_pu
-        self.ref_samecase.get_lines()[l_id_ls_ref[0]].h_pu
-                        
-                        
+        lines = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or and el.bus_ex_id in bus_ex) or (el.bus_or_id in bus_ex and el.bus_ex_id in bus_or)]
+        lines_ref = [el.id for el in self.ref_samecase.get_lines() if (el.bus_or_id in bus_or and el.bus_ex_id in bus_ex) or (el.bus_or_id in bus_ex and el.bus_ex_id in bus_or)]
+        if not lines_ref:
+            lines_ref = [el.id for el in self.ref_samecase.get_trafos() if (el.bus_hv_id in bus_or and el.bus_lv_id in bus_ex) or (el.bus_hv_id in bus_ex and el.bus_lv_id in bus_or)]
         import pdb
         pdb.set_trace()
+        # self.pp_samecase["_ppc"]["internal"]["Bbus"]
+        self.pp_samecase["_ppc"]["internal"]["Ybus"][64,67]
+        self.pp_samecase["_ppc"]["internal"]["bus"]
+                
         assert np.abs(v_ls - v_ls_ref).max() <= self.tol_eq, "error for vresults for dc"
         tmp_ = self.gridmodel.get_dcYbus() - self.ref_samecase.get_dcYbus()
         assert np.abs(tmp_).max() <= self.tol_eq, "error for dcYbus"
@@ -136,76 +139,146 @@ class AuxInitFromPyPowSyBl:
     def test_dc_pf(self):
         """test I get the same results as pandapower in dc"""
         v_ls = self.gridmodel.dc_pf(self.V_init_dc, 2, self.tol)
-        v_ls_ref = self.ref_samecase.dc_pf(self.V_init_dc, 2, self.tol)
-        assert np.abs(v_ls - v_ls_ref).max() <= self.tol_eq, "error for vresults for dc"
-        lf.run_dc(self.network)
-        pdp.rundcpp(self.pp_samecase)
-  
+        if self.compare_pp():
+            v_ls_ref = self.ref_samecase.dc_pf(self.V_init_dc, 2, self.tol)
+            assert np.abs(v_ls - v_ls_ref).max() <= self.tol_eq, "error for vresults for dc"
+        lf.run_dc(self.network_ref)
+        if self.compare_pp():
+            pdp.rundcpp(self.pp_samecase)
+        
+        # for case 300
+        # np.where(np.abs(v_ang_ls - v_ang_pypo) > 3)  # => 173, 177
+        bus_or = [173]
+        bus_ex = [177]
+        # lines = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or and el.bus_ex_id in bus_ex) or (el.bus_or_id in bus_ex and el.bus_ex_id in bus_or)]
+        lines = [el.id for el in self.gridmodel.get_lines() if (el.bus_or_id in bus_or or el.bus_or_id in bus_ex or el.bus_ex_id in bus_or or el.bus_ex_id in bus_ex)]
+        trafos = [el.id for el in self.gridmodel.get_trafos() if (el.bus_hv_id in bus_or or el.bus_hv_id in bus_ex or el.bus_lv_id in bus_or or el.bus_lv_id in bus_ex)]
+        shunts = [el.id for el in  self.gridmodel.get_shunts() if el.bus_id in bus_or or el.bus_id in bus_ex]
+        # trafo 77 => 204    2040    in cdf file
+        # trafo 85 => 
+        
         # v_mag not really relevant in dc so i study only va
-        v_ang_pypo = self.network.get_buses()["v_angle"].values
+        v_ang_pypo = self.network_ref.get_buses()["v_angle"].values
         v_ang_ls = np.rad2deg(np.angle(v_ls))
-        v_ang_pp = self.pp_samecase.res_bus["va_degree"].values
-        assert np.abs(v_ang_ls - v_ang_pp).max() <= self.tol_eq, "error for va results for dc"
-        assert np.abs(v_ang_ls - v_ang_pypo).max() <= 3 # to make the test pass... super weird !
-        # don't want to bother with the order of the lines and the lines / trafos...
+        if self.compare_pp():
+            v_ang_pp = self.pp_samecase.res_bus["va_degree"].values
+            assert np.abs(v_ang_ls - v_ang_pp).max() <= self.tol_eq, "error for va results for dc"
+        assert np.abs(v_ang_ls - v_ang_pypo).max() <= self.tol_eq
         
     def test_ac_pf(self):
         # run the powerflows
-        # v_dc = self.gridmodel.dc_pf(self.V_init_dc, 2, self.tol)
-        v_ls = self.gridmodel.ac_pf(self.V_init_ac, 10, self.tol)
-        # v_dc2 = self.ref_samecase.dc_pf(self.V_init_dc, 2, self.tol)
-        v_ls_ref = self.ref_samecase.ac_pf(self.V_init_ac, 10, self.tol)
+        v_ls = self.gridmodel.ac_pf(1.0 * self.V_init_ac, 10, self.tol)
+        if self.compare_pp():
+            v_ls_ref = self.ref_samecase.ac_pf(1.0 * self.V_init_ac, 10, self.tol)   
+            assert np.abs(v_ls - v_ls_ref).max() <= self.tol_eq, "error for vresults for ac"
+        
         param = lf.Parameters(voltage_init_mode=pp._pypowsybl.VoltageInitMode.UNIFORM_VALUES,
                               transformer_voltage_control_on=False,
                               no_generator_reactive_limits=True,
                               phase_shifter_regulation_on=False,
                               simul_shunt=False,
                               distributed_slack=False,
-                              )
-        lf.run_ac(self.network, parameters=param)
-        pdp.runpp(self.pp_samecase, init="dc")
+                              provider_parameters={"slackBusSelectionMode": "NAME",
+                                                   "slackBusesIds": self.network_ref.get_buses().iloc[self.get_slackbus_id()].name}
+                              ) 
+        res_pypow = lf.run_ac(self.network_ref, parameters=param)
+        if self.compare_pp():
+            pdp.runpp(self.pp_samecase, init="flat")
         
         # check voltage angles
-        v_ang_pypo = self.network.get_buses()["v_angle"].values
+        v_ang_pypo = self.network_ref.get_buses()["v_angle"].values
         v_ang_ls = np.rad2deg(np.angle(v_ls))
-        v_ang_pp = self.pp_samecase.res_bus["va_degree"].values
-        assert np.abs(v_ang_ls - v_ang_pp).max() <= self.tol_eq, "error for va results for ac"
-        assert np.abs(v_ang_ls - v_ang_pypo).max() <= 3 # to make the test pass... super weird !
+        import pdb
+        pdb.set_trace()
+        if self.compare_pp():
+            v_ang_pp = self.pp_samecase.res_bus["va_degree"].values - self.pp_samecase.ext_grid["va_degree"].values
+            assert np.abs(v_ang_ls - v_ang_pp).max() <= self.tol_eq, "error for va results for ac"
+        if res_pypow[0].status == pp._pypowsybl.LoadFlowComponentStatus.CONVERGED:
+            assert np.abs(v_ang_ls - v_ang_pypo).max() <= self.tol_eq
         
         # check voltage magnitudes
-        bus_ref_kv = self.network.get_voltage_levels().loc[self.network.get_buses()["voltage_level_id"].values]["nominal_v"].values
-        v_mag_pypo = self.network.get_buses()["v_mag"].values / bus_ref_kv
+        bus_ref_kv = self.network_ref.get_voltage_levels().loc[self.network_ref.get_buses()["voltage_level_id"].values]["nominal_v"].values
+        v_mag_pypo = self.network_ref.get_buses()["v_mag"].values / bus_ref_kv
         v_mag_ls = np.abs(v_ls)
-        v_mag_pp = self.pp_samecase.res_bus["vm_pu"].values
-        assert np.abs(v_mag_ls - v_mag_pp).max() <= self.tol_eq, "error for va results for dc"
-        assert np.abs(v_mag_ls - v_mag_pypo).max() <= 0.3 # to make the test pass... super weird !
+        if self.compare_pp():
+            v_mag_pp = self.pp_samecase.res_bus["vm_pu"].values
+            assert np.abs(v_mag_ls - v_mag_pp).max() <= self.tol_eq, "error for va results for dc"
+        if res_pypow[0].status == pp._pypowsybl.LoadFlowComponentStatus.CONVERGED:
+            assert np.abs(v_mag_ls - v_mag_pypo).max() <= self.tol_eq
         
         
 class TestCase14FromPypo(AuxInitFromPyPowSyBl, unittest.TestCase):
     pass
         
         
+class TestCase30FromPypo(AuxInitFromPyPowSyBl, unittest.TestCase):
+    # unittest.TestCase does not work because of https://github.com/powsybl/pypowsybl/issues/644
+    def get_pypo_grid(self):
+        return pp.network.create_ieee30()
+    
+    def get_equiv_pdp_grid(self):
+        return pn.case_ieee30()
+        
+        
+class TestCase57FromPypo(AuxInitFromPyPowSyBl, unittest.TestCase):
+    # does not appear to be the same grid !
+    def get_pypo_grid(self):
+        return pp.network.create_ieee57()
+    
+    def get_equiv_pdp_grid(self):
+        return pn.case57()
+    
+    def compare_pp(self):
+        return False
+    
+    def get_tol_eq(self):
+        return 1e-4  # otherwise vangle from pypowsybl and pandapower does not match
+                
+        
 class TestCase118FromPypo(AuxInitFromPyPowSyBl, unittest.TestCase):
+    # unittest.TestCase does not work because of https://github.com/powsybl/pypowsybl/issues/644
     def get_pypo_grid(self):
         return pp.network.create_ieee118()
+    
     def get_equiv_pdp_grid(self):
         return pn.case118()
-    def fix_grid_pypow(self):
-        vn = np.array([161.0, 138.0])
-        to_pu = vn*vn*0.01
-        self.network.update_lines(id=["L86-87-1", "L68-116-1"],
-                                  r=np.array([0.02828 , 0.00034]) * to_pu,
-                                  x=np.array([0.2074 , 0.00405]) * to_pu,
-                                  b1=np.array([0.0445, 0.164]) / 2. / to_pu,
-                                  b2=np.array([0.0445, 0.164]) / 2. / to_pu,
-                                  g1=np.array([0., 0.]),
-                                  g2=np.array([0., 0.]),
-        )
-        lf.run_ac(self.network)    
+     
     def get_gen_slack_id(self):
         return 29
+    
     def get_slackbus_id(self):
         return 68
+    
+    def get_tol_eq(self):
+        # return 3e-3  # otherwise vangle from pypowsybl and pandapower does not match
+        return 3e-5  # otherwise vangle from pypowsybl and pandapower does not match
+    
+    def compare_pp(self):
+        return False
+                
+        
+class TestCase300FromPypo(AuxInitFromPyPowSyBl, unittest.TestCase):
+        # does not work, probably grid not ordered the same
+    def get_pypo_grid(self):
+        return pp.network.create_ieee300()
+    
+    def get_equiv_pdp_grid(self):
+        return pn.case300()
+    
+    def get_tol_eq(self):
+        # return 3e-3  # otherwise vangle from pypowsybl and pandapower does not match
+        return 3e-5  # otherwise vangle from pypowsybl and pandapower does not match
+    
+    def compare_pp(self):
+        return False
+     
+    def get_gen_slack_id(self):
+        # does not work with PP, probably bus not ordered the same
+        return 55
+    
+    def get_slackbus_id(self):
+        # does not work with PP, probably bus not ordered the same
+        return 257
 
 
 if __name__ == "__main__":